backend/services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
email-validator==2.1.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.7
httpx==0.25.2
boto3==1.34.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
